+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - make `__abstract__` a constant
 - add type hints
+- rename `McsArgs.repr` to `McsArgs.qualname`
 
 ## 0.3.0 (2018/09/30)